==> 2_dataset_division.ipynb <==
import pandas as pd

interactions = pd.read_csv("./kaggle_data/interactions_train.csv")

interactions

interactions_sorted = interactions.sort_values(["u", "t"]).reset_index(drop=True)

interactions_sorted

interactions_sorted["pct_rank"] = interactions_sorted.groupby("u")["t"].rank(pct=True, method='dense')
interactions_sorted.reset_index(inplace=True, drop=True)
train_data = interactions_sorted[interactions_sorted["pct_rank"] < 0.8]
test_data = interactions_sorted[interactions_sorted["pct_rank"] >= 0.8]

print("Training set size:", train_data.shape[0])
print("Testing set size:", test_data.shape[0])

train_data.to_csv("./kaggle_data/train_data.csv", index=False)
test_data.to_csv("./kaggle_data/test_data.csv", index=False)
