==> 3_item_openai.ipynb <==
from openai import OpenAI
client = OpenAI(api_key="#####")

import pandas as pd
import json

items = pd.read_csv("./kaggle_data/items.csv")
items['Title'] = items['Title'].str.replace(r'\s*/\s*$', '', regex=True)
items

system_message = """
You are a professional book metadata assistant.

Given a JSON with:
• "title": French book title
• "subjects": list of French keywords

Return only a JSON with:

"Semantic_Tags": up to 3 concise English nouns/phrases
"Genre": 1 top-level English genre
"Subgenre": 1 specific English subgenre
"Themes": up to 2 high-level English themes
"Tone_Mood": up to 3 English descriptors (tone or style)
"Target_Audience": English phrase, ≤10 words
"Short_Summary": 1–2 English sentences, max 20 words
Output only valid JSON. No explanations or extra text.
"""

tasks = []

for index, row in items[10000:].iterrows():

    title = row["Title"]
    subjects = row["Subjects"]
    input_dict = {
        "title": title,
        "subjects": subjects
    }
    
    task = {
        "custom_id": f"task-{index}",
        "method": "POST",
        "url": "/v1/chat/completions",
        "body": {
            "model": "gpt-4.1-mini",
            "temperature": 0.1,
            "response_format": { 
                "type": "json_object"
            },
            "messages": [
                {
                    "role": "system",
                    "content": system_message
                },
                {
                    "role": "user",
                    "content": json.dumps(input_dict, ensure_ascii=False)
                }
            ],
        }
    }
    
    tasks.append(task)

file_name = "./batch_tasks_books_2.jsonl"

with open(file_name, 'w') as file:
    for obj in tasks:
        file.write(json.dumps(obj, ensure_ascii=False) + '\n')

batch_file = client.files.create(file=open(file_name, "rb"), purpose="batch")

print(batch_file)

batch_job = client.batches.create(
    input_file_id=batch_file.id,
    endpoint="/v1/chat/completions",
    completion_window="24h",
)

batch_job = client.batches.retrieve(batch_job.id)
print(batch_job)
